Renamed rewriteFiles to patchFiles and refactored
